# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cargo Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: online
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
  /stores/{storeId}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store details
      parameters:
        - in: path
          name: storeId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
  /stores/{storeId}/products:
    get:
      operationId: getStoreProducts
      tags: [stores]
      summary: Get products for a store
      parameters:
        - in: path
          name: storeId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /stores/nearby:
    get:
      operationId: getNearbyStores
      tags: [stores]
      summary: Get stores near a location
      parameters:
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
      responses:
        "200":
          description: Nearby stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"

  # ---- PRODUCTS ----
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Search and list products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: storeId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a store or product
      parameters:
        - in: query
          name: storeId
          schema:
            type: integer
        - in: query
          name: productId
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ---- FAVORITES ----
...
  /merchant/profile:
    get:
      operationId: getMerchantProfile
      tags: [stores]
      summary: Get logged-in merchant profile
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
    patch:
      operationId: updateMerchantProfile
      tags: [stores]
      summary: Update merchant profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
...
  /merchant/products:
    get:
      operationId: listMerchantProducts
      tags: [products]
      summary: List products for merchant
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /merchant/products/{productId}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ---- ADMIN: USERS ----
...
components:
  schemas:
...
    Store:
      type: object
      required: [id, name, slug, categoryName, rating, deliveryTime, deliveryFee, image, isOpen]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logo:
          type: ["string", "null"]
        image:
          type: string
        categoryName:
          type: string
        categoryIcon:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        isOpen:
          type: boolean
        isFeatured:
          type: boolean
        isVerified:
          type: boolean
        isTrending:
          type: boolean
        isOnline:
          type: boolean
        tags:
          type: array
          items:
            type: string
        distance:
          type: ["number", "null"]
    StoreDetail:
      allOf:
        - $ref: "#/components/schemas/Store"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            address:
              type: ["string", "null"]
            phone:
              type: ["string", "null"]
            whatsapp:
              type: ["string", "null"]
            latitude:
              type: ["number", "null"]
            longitude:
              type: ["number", "null"]
            openingHours:
              type: ["string", "null"]
            bannerImage:
              type: ["string", "null"]
            socialMedia:
              type: object
              properties:
                instagram:
                  type: ["string", "null"]
                facebook:
                  type: ["string", "null"]
                twitter:
                  type: ["string", "null"]
            productCategories:
              type: array
              items:
                type: string
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        address:
          type: string
        openingHours:
          type: string

    # ---- PRODUCT ----
    Product:
      type: object
      required: [id, name, price, storeId, storeName, image, categoryName]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        image:
          type: string
        storeId:
          type: integer
        storeName:
          type: string
        categoryName:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        discountPercent:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
            addons:
              type: array
              items:
                $ref: "#/components/schemas/ProductAddon"
            gallery:
              type: array
              items:
                type: string
            nutritionInfo:
              type: ["string", "null"]
            preparationTime:
              type: ["string", "null"]
    ProductVariant:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    ProductAddon:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
    ProductInput:
      type: object
      required: [name, price, categoryName]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        image:
          type: string
        categoryName:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image:
          type: string
        isAvailable:
          type: boolean

    # ---- CART ----
...
    Review:
      type: object
      required: [id, userId, userName, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        storeId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        orderId:
          type: ["integer", "null"]
        rating:
          type: number
        comment:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        storeId:
          type: integer
        productId:
          type: integer
        orderId:
          type: integer
        rating:
          type: number
        comment:
          type: string

    # ---- FAVORITES ----
...